Assets now fit graph (#50212)
Assets didn't have fit view before
added it similar to the detailsLayout page
